Logically revert "Directly show the layout selection screen."
This is logical revert of a CL [1] that was recently re-introduced as
part of my recent CL [2].

Reason for revert:
Directly showing the layout selection screen when there is no hardware
keyboard is selected yet is not something people had been familiar
with.

That behavior had been enabled only in a short period during Android N
development cycle and was never enabled in most of production builds.
To avoid confusions, let's revert back to the true Android M behavior
by logically reverting that change.

 [1]: I4483dfc89afc8d148b2cfa7c6a5f66d2a02f712a
      17b631988451613393c4656900c1821e23d8434a
 [2]: I7a2ed6dd39dcd8207d3d94e12cd01d5d67ba4bb5
      ed5b5ec60995ff4b3059ae6a1f9179e6f09a2a1c [formerly 7129b374bb57e39ac94ffd5b27dc3acc79fb1fb9]

Change-Id: If0529e20cbff432d38d9b2dd74eeb3af0da5baf1
Bug: 66498367
Fix: 75318417
Test: Manually verified

Former-commit-id: 690f103ca5d96c51f01cac2bb767a33bc6351673