SettingsRoboTests: Remove direct dep on sdk_vcurrent module.
Doesn't appear to be needed. If it was needed, the right way to depend on
this module is to specify LOCAL_SDK_VERSION. This particular module is
getting deleted soon.

RunSettingsRoboTests hangs with and without this change, so I can't really
test the runtime effects of this change, but I don't think there are any
(local_java_libraries only affects compilation).

Test: Can't test - RunSettingsRoboTests hangs with and without this change.
Bug: 77525052
Change-Id: I116cd8e9676823b6670c162aa3bbdafa516001c5
Merged-In: I9df619c3d68f0af2ee2ae944a15998275129f7c2

Former-commit-id: 97acc68065d3740b044171af7579eb2f75f56d2a