Fix line break in Opening links app list summary.
- the installed app list is using the app preference layout, which has
the appendix field to the right of the app title and summary. The
appendix and title/summary each has weight 1 to take up half of the
screen width. However, the appendix field is not used in the Opening
links page at all, so that field should be set to GONE to let the
title/summary takes up the whole available width.

Change-Id: Ia066a60dcd0a4c1d8562bd47e173378b49912d18
Fixes: 77954267
Test: visual and make RunSettingsRoboTests

Former-commit-id: 4db6f3dfb5eb594d54aa1f8ee01b7608ee81d3dc