Add Settings Search Regression test
Take a snapshot of the currently available search results, and
verify that search results aren't accidentally removed.

We use 4 items to identify a search result:
- Title
- Data Key
- Slice Uri

We use Title & Key to identify the search result, since they should not change.
The Slice Uri is used to make sure we don't regress on Slice Availability.

Test: runtest --path packages/apps/Settings/tests/unit/src/com/android/settings/search/SettingsSearchResultRegressionTest.java
Change-Id: I22498229bbcd1e90c9e0a026af9df4367a98190a

Former-commit-id: 11fa37ebb49e4ec4a975d111bf3512139b905062